Corrección models

+++ app/models/pago.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, Numeric, String, DateTime
+from sqlalchemy import Column, BigInteger, ForeignKey, Numeric, String, DateTime
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from app.db.database import Base
@@ -6,8 +6,8 @@ from app.db.database import Base
 class Pago(Base):
     __tablename__ = "pagos"
 
-    id = Column(Integer, primary_key=True, index=True)
-    reserva_id = Column(Integer, ForeignKey("reservas.id"), nullable=False)
+    id = Column(BigInteger, primary_key=True, index=True)
+    reserva_id = Column(BigInteger, ForeignKey("reservas.id"), nullable=False)
     metodo = Column(String(50), nullable=False)
     monto = Column(Numeric(10, 2), nullable=False)
     moneda = Column(String(10), default="USD")
